requirements.txt
requests
praw
six
decorator
update-checker
python-dotenv